Fixed all the issues
