+++ src/components/UserDetails.module.scss
@@ -25,16 +25,3 @@
     font-size: 1vw;
     margin-bottom: 0.7vw;
 }
-
-.newSearch {
-    background-color: black;
-    color: white;
-    font-size: 1vw;
-    padding: 0.6vw;
-    border-color: white;
-    border-width: 0.1vw;
-    border-radius: 0.3vw;
-    &:hover {
-        background: darken(white, 40%);
-    }
-}
