+++ src/components/UserMenu.tsx
@@ -8,7 +8,7 @@ import UserDetails from './UserDetails'
 import ReposList from './ReposList'
 import SingleRepo from './SingleRepo'
 
-const UserMenu = (props: {setIsUserFound: React.Dispatch<React.SetStateAction<boolean>>}) => {
+const UserMenu = ({setIsUserFound} : {setIsUserFound: React.Dispatch<React.SetStateAction<boolean>>}) => {
 
     const queryClient = useQueryClient()
     const userData = queryClient.getQueryData("user") as User
@@ -20,26 +20,38 @@ const UserMenu = (props: {setIsUserFound: React.Dispatch<React.SetStateAction<bo
         setIsRepoClicked(true)
     }
 
+    const handleClickQuit = () => {
+        setChosenRepo("")
+        setIsRepoClicked(false)
+    }
+
+    const handleNewSearch = () => {
+        queryClient.removeQueries('repos')
+        queryClient.removeQueries('user')
+        setIsUserFound(false)
+      }
+
     const {data: reposData, isSuccess, isLoading} = useGetRepos(userData.login)
     
     if(isLoading) return <h1 className={userMenuCSS.message}>Loading....</h1>
-    if (!isSuccess) return <h1 className={userMenuCSS.message}>Error in loading repos</h1>
+    if (!isSuccess) return (
+        <div>
+            <h1 className={userMenuCSS.message}>Error in loading repos</h1>
+            <button className={userMenuCSS.backBtn} onClick={handleClickQuit}> x </button>
+        </div>
+    ) 
 
     const reposNames = reposData.map((rep: Repo)=> rep.name)
 
     return (
         <div className={userMenuCSS.userMenu}>
-            <UserDetails userDetails={userData} setIsUserFound={props.setIsUserFound}/>
+            <UserDetails userDetails={userData}/>
+            <button className={userMenuCSS.newSearch} onClick={handleNewSearch}>New Search</button>
             {isRepoClicked ? 
             <div className={userMenuCSS.clickedContainer}>
                 <ReposList reposNames={reposNames} handleClickRepo={handleClickRepo}/>
                 <SingleRepo userName={userData.login} repoName={chosenRepo}/>
-                <button className={userMenuCSS.backBtn} 
-                        onClick={() => {
-                            setChosenRepo("")
-                            setIsRepoClicked(false)}}>
-                    x
-                </button>
+                <button className={userMenuCSS.backBtn} onClick={handleClickQuit}> x </button>
             </div> : 
             <div className={userMenuCSS.noClickedContainer}>
                 <ReposList reposNames={reposNames} handleClickRepo={handleClickRepo}/>
